requirements.txt
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.2
seaborn==0.13.2
scipy==1.13.1
openpyxl==3.1.5
PyYAML==6.0.2
tabulate==0.9.0
kaggle==1.6.17
transformers==4.41.2
datasets==2.20.0
accelerate==0.31.0
peft==0.11.1
trl==0.9.4
huggingface_hub>=0.33.5
torch>=2.2.0
